Crate linera_proxy

Modules

grpc 🔒

Structs

ProxyContext 🔒

ProxyOptions

Options for running the proxy.

SimpleProxy

Enums

Proxy 🔒

A Linera Proxy, either gRPC or over ‘Simple Transport’, meaning TCP or UDP. The proxy can be configured to have a gRPC ingress and egress, or a combination of TCP / UDP ingress and egress.

Statics

ALLOC 🔒

MALLOC_CONF

Configure jemalloc profiling infrastructure at startup with sampling disabled. Profiling is activated at runtime only when --enable-memory-profiling is passed.

Functions

main 🔒